add some code from Mac

+++ adapter_network.py
@@ -115,7 +115,7 @@ emb_dict =dict()
 pair_dict =dict()
 all_user_list = set(np.array(range(user_len)))
 for i in range(user_len): 
-    emb_dict[i] = np.random.rand(3584)
+    emb_dict[i] = np.random.rand(3584).astype(np.float32)
     pair_dict[i] = random.sample(all_user_list,np.random.randint(98)+2)
